Use condition variables rather than a barrier
The drawing thread is almost always slower and thus arrives at the
barrier second. Therefore, the symmetry we had before is not needed. The
driving reason for the change is that the barrier was buggy and I'm not
sure why. It would sometimes like the simulation thread through early.
There are two condition variables so that if the simulation thread does
arrive second, it will wait a full cycle before unblocking. With one
reused condition variable, the threads could get out of sync in that
situation.