Lower ESCDELAY
Curses has to wait a bit after the ESC key is read in case it is part of
an escape sequence and the terminal connection is slow. NCurses waits
one second by default, but that caused a jarring pause when ESC was
pressed. 30ms is the new wait time for all Curses implementations that
read ESCDELAY. The game is meant to be played on a fast terminal
connection, so the decreasing is unlikely to break stuff.