Always create an anonymous user
Before this change, it was possible to cause a segfault as follows:
 1. Invoke ts3d with the play_as argument to log in.
 2. Delete the user you logged in with.
 3. Go to the "Resume Game" menu.

When this occurred, the game would try to switch you to the anonymous
user, but the anonymous user would not exist since it was not created by
get_save_states called at the beginning. NULL would later be returned by
save_states_get, causing a segfault.