Allow regular compilation with only POSIX.1-2001
I like complying with standards, the older the better. This doesn't add
much value, but I wanted to do it.

The tests currently require POSIX 2008 still because of their reliance
on fmemopen and open_memstream. Changing them is not a priority, but I
will try to figure something out, maybe using pipe(2).