Add level "All Sides"
I think it might be around the difficulty of "Halls". Those two levels
might actually be harder than the ones that come after, but whatever.

"All Sides" requires you to keep track of the positions of four goblins.
This is the main challenge. It can be hard to run away from goblins
coming from all sides. I have completed this level an a 80x24 terminal.