Improve resizing responses
Resizing now works with PDCurses, although I had to write code for it
specifically. I removed the doupdate() because the code was working
without it in all implementations I tested. It even detected the resize
in the commit originally introducing the code. I just hope it wasn't
preventing some obscure crash.