Remove game management
I found that it was creating a lot of complexity with very little
benefit; why would you want multiple saves going on the same computer?
do_ts3d_game is much simpler now, and I think the UX is less confusing.
Old saves are still loaded, but are converted to the new format.